refactor: move color stuff to colors.js
